Move CliError into error module
Extract CliError into a dedicated module, adjust imports,
fix random_seed, and update tests to import RegistryError
explicitly.
